Refactor common code and tidy up
